Experiment to artificially ‘boost’ the size of our data set.
I naively tried to pair every image with every other one, producing a huge dataset. However, this very quickly becomes slow to build and so memory intensive that I’d need to build it in a different way.

However, a quick dirty solution is just to ‘randomly’ choose 10x pairings of my data set.

Testing this on MNIST data led to much better results than not doing it; the clusters are much tighter. I tried to be careful that the training and validation data sets when doing this don’t ‘bleed’ into each other, producing artificial generalization.